drm/tegra: Support Sync File-based fences in submit IOCTL
Add support for Sync File-based in-fences and out-fences to job submission.
This adds ability to pass Sync File based in-fence and get back Sync File
based out-fence during of jobs submission. A new `flags` field is used to
specify whether in-fence should be waited and out-fence created. Both
in/out-fence FD's are passed in the `fence` field.

Signed-off-by: Mikko Perttunen <mperttunen@nvidia.com>
Reviewed-by: Dmitry Osipenko <digetx@gmail.com>
Tested-by: Dmitry Osipenko <digetx@gmail.com>