drm/tegra: dc: Support legacy cursor
All Tegra generations support the so-called legacy cursor, it is a classic
2-bit color-mask cursor. Older Tegra generations (20, 30) support only the
legacy cursor, while newer can handle the ARGB8888 in addition.

There is no support of the 2-bit cursor format in DRM and generic
modesetting Xorg driver (very likely) won't ever support it too, however
we want to support it in the opentegra Xorg driver. In order to expose the
legacy cursor to userspace (opentegra), we add a custom DRM properties to
the cursor plane. The "legacy cursor" property allows Tegra DRM driver to
differentiate the modern ARGB cursor from the legacy one, still utilizing
the generic DRM cursor API. This preserves backward compatibility with the
userspace since the "legacy cursor" property is disabled by default and
userspace must enable it prior to use of the DRM cursor API.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>