drm/tegra: dc: Add minimal colorkey support for older Tegra's
Colorkey is a very useful feature supported by display controller. It is
useful for a stuff like XVideo overlay to show overlapping windows on top
of the overlay plane. The current primary user of this feature would be an
Opentegra Xorg driver which already has Xv support and colorkey is the only
major feature it lacks right now.

This patch adds support for an older Tegra SoC generations only (20,30,40),
the reason of avoiding newer generations is that the HW interface changed on
later Tegra's. It is up to somebody else to implement colorkey support for
newer generations if desired.

The colorkey is coupled with plane blending controls. The blending controls
are quite beefy and right now we are not interested in their support, so we
add a minimal colorkey support that is sufficient for Xv implementation
and is future-proof in case we'd want to add blending controls later. In
order to expose colorkey controls to userspace, we use DRM atomic properties.
The controls are: colorkey enable and the actual colorkey ABGR values.
The colorkey 'enable' property is attached to CRTC and the colors to the
planes, following the display controller HW topology. Since we do not
support the blending controls, attaching color properties to a plane other
than primary doesn't make sense right now, because the current-hardcoded
blending configuration isn't useful for the overlay planes colorkey'ing.
However the hardcoded blending config fits neatly for an Xv use-case,
enabling a colorkey for the primary plane causes the pixels not matching
the color key to be drawn on top of overlay planes.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>