drm/tegra: Introduce caching of IOVA mappings for Tegra20
On Tegra20 due to a small size of GART's aperture we can't have a
persistent IOVA mapping for the BO. We solve this by mapping BO's IOVA
dynamically on its pinning for a job and unmapping it on the unpinning.
In most cases the same BO would be constantly remapped, this is quite
inefficient and results in around 30% CPU usage overhead for OpenTegra
Xorg driver. To solve this inefficiency this patch introduces a caching
of IOVA mappings for Tegra20, newer generations use SMMU instead of GART
and do not require it. The cache is implemented using generic DRM memory
management 'scanning' mechanism. Now, instead of unmapping pages on BO
unpinning, this mapping is put on a 'eviction' list and there are three
variants for a BO pinning:

1) The current BO's mapping is already on the 'eviction' list and we
   simply remove it from the list and return, that's very efficient.

2) We try a generic IOVA allocation and return on its success.

3) In case of out-of-IOVA-space, the DRM's mm scanner would search
   through the eviction list for a suitable hole for this allocation and
   the 'victim' mappings would be released from the cache, resulting in
   a success on the allocation retry.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>