drm/tegra: Implement buffer objects reservation
Before submitting a job to HW, we need to ensure that none of the job BOs
are currently in use by some other device, otherwise we may need to wait
until these BOs become available for use. That's what DMABUF reservation
solves for us. In a case of write to BO, we need to wait for all reads and
writes to be completed first, a BO relocation flag is required to designate
BO as writable, so we extend the Tegra DRM UAPI with a new BO relocation
flag - 'BO read'. A 'read' flag is preferred over 'write' to preserve
compatibility with older kernels, since relocation is designated as
writable by default and write has stronger ordering requirement than read.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>