drm/tegra: Introduce "CPU-prepare" IOCTL
The new IOCTL does one very useful thing, it waits for a completion of
all outstanding HW operations for a given BO in a given time. A 'BO write'
flag is used to designate type of fences to wait for (shared or exclusive.)
Right now this new IOCTL is primarily required for implementing BO caching
in libdrm-tegra.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>