gpu: host1x: Reset clients HW on CDMA timeout
In case of a HW hang all further job submissions would fail, currently
the only way to recover is a reboot. The hung HW unit needs to be reset
in order to proceed without a reboot. Actually there is a comment in the
code suggesting that cdma_freeze() would reset host1x module, but there
is no code implementing that reset, let's implement it.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>