stmmac: fix breakage in stmmac_hw_setup()
Since "drivers: net: stmmac: reworking the PCS code" ->pcs_ctrl_ane()
had been taking iomem address to access as the first argument; its
predecessor (->ctrl_ane()) used to take struct mac_device_info instead.

One of the callers had not been converted; as the result, instead of
reading and modifying a word in card iomem we read and modify a word
in (or near) the in-core strucct mac_device_info.

Fixes: 70523e639bf8 (drivers: net: stmmac: reworking the PCS code)
Cc: stable@vger.kernel.org
Signed-off-by: Al Viro <viro@zeniv.linux.org.uk>