gpu: host1x: Remove unneeded NULL-check of waiter
Waiter won't ever be NULL unless there is some sever bug in the code.
It looks like that it is some kind of a leftover debug checking that
isn't really needed. Let's remove it to cleanup code a tad. Now that
this function never fails, let's change its return type to void to
cleanup code a tad.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>