gpu: host1x: Support HW context switching
In order to perform a HW context switch the following steps should be
accomplished:

1) Check whether context store is required and schedule a context store
   work if yes.

2) Check whether context restore is required and prepend job with a context
   restore gather if yes.

Context store is implemented using indirect module read work that is
issued on a syncpoint interrupt in case of a 'software' store.

The 'store' work performs fetching of indirectly read data from channels
output FIFO and writes it to contexts BO.

Context restore is implemented using a CDMA gather that reads context data
from contexts BO and writes it to client's module.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>