dax: disable filesystem dax on devices that do not map pages
If a dax buffer from a device that does not map pages is passed to read(2)
or write(2) as a target for direct-I/O it triggers SIGBUS.  If gdb
attempts to examine the contents of a dax buffer from a device that does
not map pages it triggers SIGBUS.  If fork(2) is called on a process with
a dax mapping from a device that does not map pages it triggers SIGBUS.
'struct page' is required otherwise several kernel code paths break in
surprising ways.  Disable filesystem-dax on devices that do not map pages.

Link: http://lkml.kernel.org/r/150664807247.36094.11168730579639072446.stgit@dwillia2-desk3.amr.corp.intel.com
Signed-off-by: Dan Williams <dan.j.williams@intel.com>
Cc: Jan Kara <jack@suse.cz>
Cc: Jeff Moyer <jmoyer@redhat.com>
Cc: Christoph Hellwig <hch@lst.de>
Cc: Ross Zwisler <ross.zwisler@linux.intel.com>
Cc: "Kirill A. Shutemov" <kirill.shutemov@linux.intel.com>
Cc: Michal Hocko <mhocko@suse.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>
Signed-off-by: Mark Brown <broonie@kernel.org>