drm: Introduce exclusive syncpoint
In order to perform a cross-channels synchronization, we need to insert
waitchecks into the commands stream, these waitchecks will wait for
syncpoint to reach a specified threshold value. In order to setup
thresholds we need to know the exact syncpoint values, so userspace
need an ability to request an exclusive syncpoint that isn't used by
any other context.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>