gpu: host1x: Increment syncpoint after submitting gathers
In order to serialize jobs properly, the newly submitted job should wait
for the previous to be completed. Currently we are relying on userspace
to append job in the end with a syncpoint increment that signals jobs
completion, let's not rely on userspace and append the submitted job
with the increment in the kernel driver.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>