clocksource/drivers/spreadtrum: Add timer driver for Spreadtrum SC9860 platform
The Spreadtrum SC9860 platform will use the architected timers as local
clock events, but we also need a broadcast timer device to wakeup the
cpus when the cpus are in sleep mode.

The Spreadtrum timer can support 32bit or 64bit counter, as well as
supporting period mode or one-shot mode.

Signed-off-by: Baolin Wang <baolin.wang@spreadtrum.com>
Acked-by: Philippe Ombredanne <pombredanne@nexb.com>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>