mmc: core: Add raw_boot_mult field to mmc_ext_csd
Bootloader of NVIDIA Tegra devices linearizes the boot0/boot1/main
partitions into a single virtual space, and thus, all partition addresses
are shifted by the size of boot0 + boot1 partitions.  The offset needs to
be known in order to find the EFI entry on internal EMMC storage of Tegra
devices.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>