ext4: reject mount options not supported when remounting in handle_mount_opt()
Rejecting the mount options in ext4_remount() means that some mount
options would be enabled for a small amount of time, and then the
mount option change would be reverted.  In the case of "mount -o
remount,dax", this can cause a race where files would temporarily
treated as DAX --- and then not.

Cc: stable@kernel.org
Reported-and-tested-by: syzbot+bca9799bf129256190da@syzkaller.appspotmail.com
Signed-off-by: Theodore Ts'o <tytso@mit.edu>
Link: https://lore.kernel.org/r/20200415174839.461347-1-tytso@mit.edu
Signed-off-by: Theodore Ts'o <tytso@mit.edu>