XXX: WIP: drm: Universal display rotation on Nexus 7
This removes a need to have bleeding edge userspace that knows how to
work with the rotated planes. Only Opentegra driver supports planes rotation
today, everything else is upside-down without the universal rotation support.

The plan is to teach DRM core about the CRTC rotation capabilities and then
this hack won't be needed, for now it's not obvious how to implement that.

Signed-off-by: Dmitry Osipenko <digetx@gmail.com>