io_uring: set ctx need-wakeup flag when SQPOLL thread is going idle
Don't set the ctx SQPOLL need-wakeup flag before we're going idle for all
rings in this shared mapping. We could have some rings that are idle, but
also still have ones that are either spinning or processing work. This
can cause extraneous system calls for the rings that are idle, while the
thread is still doing work.

Fixes: eeb9711daf2c ("io_uring: enable IORING_SETUP_ATTACH_WQ to attach to SQPOLL thread too")
Reported-by: Xiaoguang Wang <xiaoguang.wang@linux.alibaba.com>
Signed-off-by: Jens Axboe <axboe@kernel.dk>