rcu/tree: Use schedule_delayed_work() instead of WQ_HIGHPRI queue
Recently the separate worker thread has been introduced to
maintain the local page cache from the regular kernel context,
instead of kvfree_rcu() contexts. That was done because a caller
of the k[v]free_rcu() can be any context type what is a problem
from the allocation point of view.

On the other hand, the lock-less way of obtaining a page has been
introduced and directly injected to the k[v]free_rcu() path.

Therefore it is not important anymore to use a high priority "wq"
for the external job that used to fill a page cache ASAP when it
was empty.

Signed-off-by: Uladzislau Rezki (Sony) <urezki@gmail.com>
[ paulmck: Add rcu_scheduler_active check to avoid early boot hangs. ]
Signed-off-by: Paul E. McKenney <paulmck@kernel.org>